Add error to lickdir; change size_t to int
Add error field to lickdir_t
Add new_lick_dir and free_lick_dir
Change a size_t in llist to an int

+++ liblick/llist.c
@@ -9,8 +9,8 @@ node_t *new_node(void *val, node_t *next) {
     return n;
 }
 
-size_t list_length(node_t *lst) {
-    size_t len = 0;
+int list_length(node_t *lst) {
+    int len = 0;
     while(lst != NULL) {
         len++;
         lst = lst->next;
